Update and rename working-solution-P.py to solution2-nc.py
file name changed
Hidden test cases updated
When there is a valid path before removing was not counted in paths list in the previous solution. Foobar also didn't complain about this. Luckily spotted now and updated the code. 
This code works. It gives the path, all possible paths, and the shortest path.